pyproject.toml
[tool.poetry]
name = "shorts-automation-project"
version = "0.1.0"
description = ""
authors = ["Seth Mcfeeters <seth.mcfeeters@gmail.com>"]
readme = "README.md"
packages = [{include = "shorts_automation_project"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
moviepy = "^1.0.3"
pillow = "^9.5.0"
openai = "^0.27.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
